Moved endpoint to RegistrationController
