+++ backend/src/main/java/com/sumotournamenthub/backend/controller/AgeCategoryController.java
@@ -47,10 +47,4 @@ public class AgeCategoryController {
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
-    @GetMapping("/athlete/{athleteId}/season/{seasonId}")
-    public ResponseEntity<AgeCategoryDto> getAgeCategoryByAthleteIdAndSeasonId(@PathVariable Integer athleteId, @PathVariable Integer seasonId) {
-        var ageCategory = ageCategoryService.getAgeCategoryByAthleteIdAndSeasonId(athleteId, seasonId);
-        return new ResponseEntity<>(ageCategory, HttpStatus.OK);
-    }
-
 }
