+++ backend/src/main/java/com/sumotournamenthub/backend/controller/RegistrationController.java
@@ -1,5 +1,6 @@
 package com.sumotournamenthub.backend.controller;
 
+import com.sumotournamenthub.backend.dto.AgeCategoryDto;
 import com.sumotournamenthub.backend.dto.RegistrationDto;
 import com.sumotournamenthub.backend.service.RegistrationService;
 import org.springframework.http.HttpStatus;
@@ -20,6 +21,12 @@ public class RegistrationController {
     public ResponseEntity<RegistrationDto> registerAthlete(@RequestBody RegistrationDto dto) {
         var registration = service.registerAthleteForWeightCategory(dto);
         return new ResponseEntity<>(registration, HttpStatus.CREATED);
+    }
+
+    @GetMapping("/athlete/{athleteId}/season/{seasonId}")
+    public ResponseEntity<AgeCategoryDto> getAgeCategoryByAthleteIdAndSeasonId(@PathVariable Integer athleteId, @PathVariable Integer seasonId) {
+        var ageCategory = service.getAgeCategoryByAthleteIdAndSeasonId(athleteId, seasonId);
+        return new ResponseEntity<>(ageCategory, HttpStatus.OK);
     }
 
 }
