+++ backend/src/main/java/com/sumotournamenthub/backend/service/AgeCategoryService.java
@@ -16,8 +16,6 @@ import java.util.List;
 @RequiredArgsConstructor
 public class AgeCategoryService {
     private final AgeCategoryRepository repository;
-    private final AthleteService athleteService;
-    private final SeasonService seasonService;
 
     public AgeCategoryDto createAgeCategory(AgeCategoryDto ageCategoryDto) {
         AgeCategory newAgeCategory = createOrRetrieveCategory(ageCategoryDto);
@@ -56,15 +54,6 @@ public class AgeCategoryService {
         return repository.findByNameAndAgeLowerBoundAndAgeUpperBoundAndGender(
                         dto.getAgeCategoryName(), dto.getAgeLowerBound(), dto.getAgeUpperBound(), dto.getGender())
                 .orElseGet(() -> new AgeCategory(dto.getAgeCategoryName(), dto.getAgeLowerBound(), dto.getAgeUpperBound(), dto.getGender()));
-    }
-
-    public AgeCategoryDto getAgeCategoryByAthleteIdAndSeasonId(Integer athleteId, Integer seasonId) {
-        var athleteAgeCategories = athleteService.getAgeCategoryByAthleteId(athleteId);
-        var seasonAgeCategories = seasonService.getAllAgeCategories(seasonId);
-        return seasonAgeCategories.stream()
-                .filter(athleteAgeCategories::contains)
-                .findFirst()
-                .orElse(null);
     }
 
     public AgeCategoryDto convertToDto(AgeCategory ageCategory) {
