+++ backend/src/main/java/com/sumotournamenthub/backend/service/AthleteService.java
@@ -1,12 +1,10 @@
 package com.sumotournamenthub.backend.service;
 
 import com.sumotournamenthub.backend.domain.Athlete;
-import com.sumotournamenthub.backend.dto.AgeCategoryDto;
 import com.sumotournamenthub.backend.dto.AthleteDto;
 import com.sumotournamenthub.backend.repository.AthleteRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import java.util.List;
 
 import static com.sumotournamenthub.backend.utils.ExceptionUtils.entityNotFound;
 
@@ -16,7 +14,6 @@ public class AthleteService {
 
     private final AthleteRepository repository;
     private final ClubService clubService;
-    private final RegistrationService registrationService;
 
     public AthleteDto createAthlete(AthleteDto dto) {
         var club = clubService.getClubEntity(dto.getClubId());
@@ -32,12 +29,6 @@ public class AthleteService {
 
     public Athlete getAthleteEntity(int id) {
         return repository.findById(id).orElseThrow(() -> entityNotFound("Athlete", id));
-    }
-
-    public List<AgeCategoryDto> getAgeCategoryByAthleteId(int athleteId) {
-        var athlete = getAthleteEntity(athleteId);
-        return athlete.getRegistrations().stream()
-                .map(registration -> registrationService.getAgeCategoryByRegistrationId(registration.getId())).toList();
     }
 
     public static AthleteDto convertToDto(Athlete athlete) {
